REF: Factor out getting local `octue` version
skipci
