+++ octue/cloud/__init__.py
@@ -1,12 +1,6 @@
-import importlib.metadata
-
 import octue.exceptions
-import twined.exceptions
 from octue.utils.exceptions import create_exceptions_mapping
-
-
-LOCAL_SDK_VERSION = importlib.metadata.version("octue")
-
+import twined.exceptions
 
 EXCEPTIONS_MAPPING = create_exceptions_mapping(
     globals()["__builtins__"],
